% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiomics.R,
%   R/Multiview_Learner.R
\name{Multiview_learner}
\alias{Multiview_learner}
\title{Simulated Multi-Omics Data Generation}
\usage{
Multiview_learner(
  mainDir = "/Users/siddheshkulkarni/Desktop",
  subDir = "plots_simulated_data",
  seed.number = 1234,
  M = 3,
  p_m = c(500, 700, 1000),
  n = 100,
  nTest = 100,
  params = list(fixed_act_pattern = TRUE, K0 = 10, nPatternJ = 5, J0_m = c(5, 6, 7), K0_m
    = c(5, 7, 10), S0_m = c(6, 7, 8), nu2_0 = 0.1, pi2_0 = 0.1, snr_m = NULL, snr_y = 1,
    decaying_loadings = FALSE, decaying_type = "sqrt", pJ_zero = 0.6, pJ_sign = 0.4,
    pJ_zeroEntr = 0.5, pS_zero = 0.6, pS_sign = 0.4, pS_zeroEntr = 0.5),
  Activity_pattern_plot = TRUE,
  Emprical_correlation_plot = TRUE,
  Crossprd_plot = TRUE,
  Data_save = TRUE
)

Multiview_learner(
  mainDir = "/Users/siddheshkulkarni/Desktop",
  subDir = "plots_simulated_data",
  seed.number = 1234,
  M = 3,
  p_m = c(500, 700, 1000),
  n = 100,
  nTest = 100,
  params = list(fixed_act_pattern = TRUE, K0 = 10, nPatternJ = 5, J0_m = c(5, 6, 7), K0_m
    = c(5, 7, 10), S0_m = c(6, 7, 8), nu2_0 = 0.1, pi2_0 = 0.1, snr_m = NULL, snr_y = 1,
    decaying_loadings = FALSE, decaying_type = "sqrt", pJ_zero = 0.6, pJ_sign = 0.4,
    pJ_zeroEntr = 0.5, pS_zero = 0.6, pS_sign = 0.4, pS_zeroEntr = 0.5),
  Activity_pattern_plot = TRUE,
  Emprical_correlation_plot = TRUE,
  Crossprd_plot = TRUE,
  Data_save = TRUE
)
}
\arguments{
\item{mainDir}{Character. The main directory where plots and data will be saved.}

\item{subDir}{Character. The subdirectory within \code{mainDir} for saving plots and data.}

\item{seed.number}{Numeric. Seed for random number generation.}

\item{M}{Integer. Number of views.}

\item{p_m}{Integer vector. Number of features per view.}

\item{n}{Integer. Train set size.}

\item{nTest}{Integer. Test set size.}

\item{params}{List. A list of parameters including:
\itemize{
\item \code{fixed_act_pattern} Logical. Whether to use a fixed activity pattern.
\item \code{K0} Integer. Number of factors.
\item \code{nPatternJ} Integer. Number of joint patterns.
\item \code{J0_m} Integer vector. Number of groups in the joint component per view.
\item \code{K0_m} Integer vector. Number of factors in view-specific components.
\item \code{S0_m} Integer vector. Number of groups in view-specific components.
\item \code{nu2_0} Numeric. Variability in the joint loadings.
\item \code{pi2_0} Numeric. Variability in the view-specific loadings.
\item \code{snr_m} Numeric vector. Signal-to-noise ratio for each view.
\item \code{snr_y} Numeric. Signal-to-noise ratio for the response.
\item \code{decaying_loadings} Logical. Whether to enforce loadings decay.
\item \code{decaying_type} Character. Type of decay ('sqrt' or 'log').
\item \code{pJ_zero} Numeric. Probability of group-wise zero in joint component.
\item \code{pJ_sign} Numeric. Probability of group-wise sign-switch in joint component.
\item \code{pJ_zeroEntr} Numeric. Probability of entry-wise zero in joint component.
\item \code{pS_zero} Numeric. Probability of group-wise zero in specific components.
\item \code{pS_sign} Numeric. Probability of group-wise sign-switch in specific components.
\item \code{pS_zeroEntr} Numeric. Probability of entry-wise zero in specific components.
}}

\item{Activity_pattern_plot}{Logical. Whether to save the activity pattern plot.}

\item{Emprical_correlation_plot}{Logical. Whether to save the empirical correlation plot.}

\item{Crossprd_plot}{Logical. Whether to save crossproduct plots.}

\item{Data_save}{Logical. Whether to save the generated data.}

\item{fixed_act_patter}{Logical. Set a custom activity pattern in the joint component. Default is TRUE.}

\item{J0_m}{Integer vector. Number of groups in the joint component. Default is c(5, 6, 7).}

\item{K0_m}{Integer vector. Number of factors in view-specific components. Default is c(5, 7, 10).}

\item{S0_m}{Integer vector. Number of groups in view-specific components. Default is c(6, 7, 8).}

\item{nu2_0}{Numeric. Input the joint loadings variability. Default is 0.1.}

\item{pi2_0}{Numeric. Input the view-specific loadings variability. Default is 0.1.}

\item{decaying_loadings}{Logical. Enforce loadings decay (re-weighting importance of latent axis of variation). Default is FALSE.}

\item{decaying_type}{Character. Type of decay ('sqrt' or 'log'). Default is 'sqrt'.}

\item{pJ_zero}{Numeric. Probability of group-wise zero in joint component. Default is 0.6.}

\item{pJ_sign}{Numeric. Probability of group-wise sign-switch in joint component. Default is 0.4.}

\item{pJ_zeroEntr}{Numeric. Probability of entry-wise zero in joint component. Default is 0.5.}

\item{pS_zero}{Numeric. Probability of group-wise zero in specific components. Default is 0.6.}

\item{pS_sign}{Numeric. Probability of group-wise sign-switch in specific components. Default is 0.4.}

\item{pS_zeroEntr}{Numeric. Probability of entry-wise zero in specific components. Default is 0.5.}
}
\value{
List containing simulated multi-omics data and parameters.

A list containing the generated data and various parameters used in the simulation.
}
\description{
This function generates simulated multi-omics data for benchmarking and testing.

This function generates synthetic multi-view data and performs various analyses, including plotting correlations and activity patterns.
}
\examples{
mainDir <- tempdir()
subDir <- 'plots_simulated_data'
seed.number <- 1234
M <- 2
p_m <- c(500, 700)
n <- 100
nTest <- 100
params <- list(
  fixed_act_pattern = TRUE,
  K0 = 10,
  nPatternJ = 5,
  J0_m = c(4, 5),
  K0_m = c(5, 6),
  S0_m = c(4, 5),
  nu2_0 = 0.1,
  pi2_0 = 0.1,
  snr_m = NULL,
  snr_y = 1,
  decaying_loadings = FALSE,
  decaying_type = 'sqrt',
  pJ_zero = 0.6,
  pJ_sign = 0.4,
  pJ_zeroEntr = 0.5,
  pS_zero = 0.6,
  pS_sign = 0.4,
  pS_zeroEntr = 0.5
)
Activity_pattern_plot <- TRUE
Emprical_correlation_plot <- TRUE
Crossprd_plot <- TRUE
Data_save <- TRUE
Data <- Multiview_learner(mainDir, subDir, seed.number, M, p_m, n, nTest, params, Activity_pattern_plot, Emprical_correlation_plot, Crossprd_plot, Data_save)

}
